Reworked 4.0, again.
