+++ src/dgbowl_schemas/yadg/__init__.py
@@ -3,8 +3,6 @@ from typing import Union
 from .dataschema_4_1 import DataSchema as DataSchema_4_1
 from .dataschema_4_0 import DataSchema as DataSchema_4_0
 
-DataSchema = DataSchema_4_1
-
 def to_dataschema(**kwargs):
     DataSchemas = Union[DataSchema_4_1, DataSchema_4_0]
     return parse_obj_as(DataSchemas, kwargs)
