+++ src/dgbowl_schemas/yadg/dataschema_4_0/__init__.py
@@ -1,9 +1,9 @@
-from pydantic import BaseModel, Extra
+from pydantic import BaseModel, Extra, Field
 from typing import Sequence
 from .metadata import Metadata
-from .step import Step
+from .step import Steps
 
 
 class DataSchema(BaseModel, extra=Extra.forbid):
     metadata: Metadata
-    steps: Sequence[Step]
+    steps: Sequence[Steps]
