+++ src/dgbowl_schemas/yadg/dataschema_4_0/parameters.py
@@ -1,0 +1,11 @@
+from pydantic import BaseModel, Extra, Field
+from typing import Literal, Optional, Union, Any, Mapping
+
+from .timestamp import Timestamp, TimeDate, UTS
+
+
+class Tol(BaseModel, extra=Extra.forbid):
+    atol: Optional[float]
+    rtol: Optional[float]
+
+Timestamps = Union[Timestamp, TimeDate, UTS]
