+++ src/dgbowl_schemas/yadg/dataschema_4_1/externaldate.py
