--- src/dgbowl_schemas/yadg_dataschema/dataschema_4_1/__init__.py
@@ -1,1 +1,0 @@
-from .schema import DataSchema
