+++ tests/test_dataschema.py
@@ -2,7 +2,7 @@ import pytest
 import os
 import yaml
 import json
-from dgbowl_schemas.yadg_dataschema import to_dataschema
+from dgbowl_schemas.yadg import to_dataschema
 
 from ref_dataschema import ts0, ts1, ts2, ts3, ts4, ts5
